[#82] Campo Grade permite notas fora do intervalo 0.0 - 10.0, e backend não aceita float (#120)
* feat: change grade type from Integer to Float across models and update related logic

* fix: change migration name

* feat: change grade type from Float to BigDecimal across models and update related logic

* fix: correct validation message for grade input

---------

+++ server/src/main/java/br/ufal/ic/odontolog/dtos/ReviewDTO.java
@@ -2,13 +2,14 @@ package br.ufal.ic.odontolog.dtos;
 
 import br.ufal.ic.odontolog.enums.ReviewStatus;
 import br.ufal.ic.odontolog.models.Supervisor;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class ReviewDTO {
   private Long id;
   private String comments;
-  private Integer grade;
+  private BigDecimal grade;
   private ReviewStatus reviewStatus;
   private Supervisor supervisor;
 }
