+++ server/src/main/java/br/ufal/ic/odontolog/dtos/ReviewableSubmitSupervisorReviewDTO.java
@@ -1,13 +1,14 @@
 package br.ufal.ic.odontolog.dtos;
 
 import jakarta.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
 public class ReviewableSubmitSupervisorReviewDTO {
   @NotNull private String comments;
 
-  @NotNull private Integer grade;
+  @NotNull private BigDecimal grade;
 
   @NotNull private Boolean approved;
 }
